Move flame graph to its own component
