+++ frontend/src/pages/Traces/TracePage.tsx
@@ -1,31 +1,19 @@
-import {
-	Badge,
-	Box,
-	ButtonIcon,
-	Heading,
-	IconSolidZoomIn,
-	IconSolidZoomOut,
-	Stack,
-	Tabs,
-	Text,
-} from '@highlight-run/ui'
+import { Badge, Box, Heading, Stack, Tabs, Text } from '@highlight-run/ui'
 import moment from 'moment'
 import React, { useMemo, useState } from 'react'
 
 import LoadingBox from '@/components/LoadingBox'
 import { useGetTraceQuery } from '@/graph/generated/hooks'
-import { Trace, TraceError } from '@/graph/generated/schemas'
+import { TraceError } from '@/graph/generated/schemas'
 import { TraceErrors } from '@/pages/Traces/TraceErrors'
-import { TraceFlameGraphNode } from '@/pages/Traces/TraceFlameGraphNode'
+import { TraceFlameGraph } from '@/pages/Traces/TraceFlameGraph'
 import { TraceLogs } from '@/pages/Traces/TraceLogs'
 import { TraceSpanAttributes } from '@/pages/Traces/TraceSpanAttributes'
 import {
 	FlameGraphSpan,
 	getFirstSpan,
-	getMaxDepth,
 	getTraceDurationString,
 	getTraceTimes,
-	organizeSpans,
 } from '@/pages/Traces/utils'
 import { useParams } from '@/util/react-router/useParams'
 
@@ -39,11 +27,6 @@ enum TraceTabs {
 
 type Props = {}
 
-const MAX_TICKS = 6
-export const ticksHeight = 24
-export const outsidePadding = 4
-export const lineHeight = 18
-
 export const TracePage: React.FC<Props> = () => {
 	const {
 		project_id: projectId,
@@ -54,19 +37,10 @@ export const TracePage: React.FC<Props> = () => {
 		trace_id: string
 		span_id?: string
 	}>()
-	const svgContainerRef = React.useRef<HTMLDivElement>(null)
 	const [activeTab, setActiveTab] = useState<TraceTabs>(TraceTabs.Info)
-	const [selectedSpan, setSelectedSpan] = useState<FlameGraphSpan | Trace>()
 	const [hoveredSpan, setHoveredSpan] = useState<FlameGraphSpan>()
-	const [tooltipCoordinates, setTooltipCoordinates] = useState({
-		x: 0,
-		y: 0,
-	})
-	const [zoom, setZoom] = useState(1)
+	const [selectedSpan, setSelectedSpan] = useState<FlameGraphSpan>()
 	const highlightedSpan = hoveredSpan || selectedSpan
-
-	// TODO: Make dynamic. Consider using auto sizer.
-	const width = 660
 
 	const { data, loading } = useGetTraceQuery({
 		variables: {
@@ -75,41 +49,23 @@ export const TracePage: React.FC<Props> = () => {
 		},
 		onCompleted: (data) => {
 			if (spanId) {
-				const span = data.trace?.trace.find(
+				const selectedSpan = data.trace?.trace.find(
 					(span) => span.spanID === spanId,
 				)
-				if (span) {
-					setSelectedSpan(span)
+
+				if (selectedSpan) {
+					setSelectedSpan(selectedSpan as FlameGraphSpan)
+				} else {
+					const rootSpan = getFirstSpan(data.trace?.trace ?? [])
+
+					if (rootSpan) {
+						setSelectedSpan(rootSpan as FlameGraphSpan)
+					}
 				}
 			}
 		},
 		skip: !projectId || !traceId,
 	})
-
-	const setTooltipCoordinatesImpl = React.useCallback(
-		(e: React.MouseEvent) => {
-			const elementBounds = e.currentTarget.getBoundingClientRect()
-			const y = elementBounds.top - 60
-			const x = e.clientX
-
-			setTooltipCoordinates({ x, y })
-		},
-		[],
-	)
-
-	const traces = React.useMemo(() => {
-		if (!data?.trace?.trace) return []
-		const sortableTraces = [...data.trace.trace]
-
-		if (!selectedSpan) {
-			const firstSpan = getFirstSpan(sortableTraces)
-			setSelectedSpan(firstSpan)
-		}
-
-		// TODO: Figure out how we're ending up with multiple root spans...
-		return organizeSpans(sortableTraces)
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [data?.trace])
 
 	const { startTime, totalDuration } = useMemo(() => {
 		if (!data?.trace) {
@@ -132,14 +88,6 @@ export const TracePage: React.FC<Props> = () => {
 		[firstSpan],
 	)
 
-	const height = useMemo(() => {
-		if (!traces.length) return 260
-
-		const maxDepth = getMaxDepth(traces)
-		const lineHeightWithPadding = lineHeight + 4
-		return maxDepth * lineHeightWithPadding + ticksHeight + outsidePadding
-	}, [traces])
-
 	const errors = useMemo(() => {
 		if (!data?.trace?.errors) {
 			return [] as TraceError[]
@@ -149,22 +97,6 @@ export const TracePage: React.FC<Props> = () => {
 	}, [data?.trace?.errors])
 
 	const durationString = getTraceDurationString(totalDuration)
-	const ticks = useMemo(() => {
-		if (!totalDuration) return []
-
-		const length = Math.round(MAX_TICKS * zoom)
-		return Array.from({ length }).map((_, index) => {
-			const percent = index / (length - 1)
-			const tickDuration = totalDuration * percent
-			const time = getTraceDurationString(tickDuration)
-
-			return {
-				time: time.trim() === '' ? '0ms' : time,
-				percent,
-				x: width * percent * zoom,
-			}
-		})
-	}, [totalDuration, zoom, width])
 
 	if (!data?.trace || !data?.trace?.trace?.length) {
 		return loading ? (
@@ -174,32 +106,6 @@ export const TracePage: React.FC<Props> = () => {
 				<Text>Trace not found</Text>
 			</Box>
 		)
-	}
-
-	const handleZoom = (
-		e: React.MouseEvent<HTMLButtonElement, MouseEvent>,
-		additionalZoom: number,
-	) => {
-		e.preventDefault()
-		e.stopPropagation()
-
-		const newZoom = Math.max(1, Math.round(zoom + additionalZoom))
-		const currentScrollPosition = svgContainerRef.current?.scrollLeft ?? 0
-		const visibleCenter = currentScrollPosition + width / 2
-		const newVisibleCenter = visibleCenter * (newZoom / zoom)
-		const isScrolledAllTheWayToTheLeft = currentScrollPosition <= 1
-		const isScrolledAllTheWayToTheRight =
-			currentScrollPosition >= width * zoom - (width + 2)
-		const newScrollPosition = isScrolledAllTheWayToTheRight
-			? width * newZoom - width
-			: isScrolledAllTheWayToTheLeft
-			? currentScrollPosition
-			: newVisibleCenter - width / 2
-
-		setZoom(newZoom)
-		setTimeout(() => {
-			svgContainerRef.current?.scrollTo(newScrollPosition, 0)
-		})
 	}
 
 	return (
@@ -220,155 +126,16 @@ export const TracePage: React.FC<Props> = () => {
 					</Box>
 				</Stack>
 
-				<Box
-					backgroundColor="raised"
-					borderRadius="6"
-					border="dividerWeak"
-					position="relative"
-				>
-					<Box
-						display="flex"
-						gap="2"
-						position="absolute"
-						style={{
-							bottom: 4,
-							left: 4,
-						}}
-					>
-						<ButtonIcon
-							size="xSmall"
-							kind="secondary"
-							icon={<IconSolidZoomIn />}
-							onClick={(e) => handleZoom(e, 1)}
-						>
-							Zoom In
-						</ButtonIcon>
-						<ButtonIcon
-							size="xSmall"
-							kind="secondary"
-							icon={<IconSolidZoomOut />}
-							onClick={(e) => handleZoom(e, -1)}
-						>
-							Zoom Out
-						</ButtonIcon>
-					</Box>
-					<Box
-						ref={svgContainerRef}
-						overflowY="scroll"
-						style={{
-							maxHeight: 300,
-						}}
-					>
-						<svg
-							xmlns="http://www.w3.org/2000/svg"
-							height={height}
-							width={width * zoom}
-							style={{ display: 'block' }}
-						>
-							<line
-								stroke="#e4e2e4"
-								x1={0}
-								y1={ticksHeight}
-								x2={width * zoom}
-								y2={ticksHeight}
-							/>
-
-							{ticks.map((tick) => {
-								const isFirstTick = tick.percent === 0
-								const isLastTick = tick.percent === 1
-								const x = isFirstTick
-									? outsidePadding
-									: isLastTick
-									? tick.x - outsidePadding
-									: tick.x
-
-								return (
-									<g key={tick.time}>
-										<line
-											x1={x}
-											y1={ticksHeight - 8}
-											x2={x}
-											y2={ticksHeight - 2}
-											stroke="#e4e2e4"
-										/>
-
-										<text
-											x={x}
-											y={12}
-											fill="#6f6e77"
-											fontSize={10}
-											textAnchor={
-												isFirstTick
-													? 'start'
-													: isLastTick
-													? 'end'
-													: 'middle'
-											}
-										>
-											{tick.time}
-										</text>
-									</g>
-								)
-							})}
-							{traces.map((span, index) => {
-								return (
-									<TraceFlameGraphNode
-										key={index}
-										errors={errors}
-										span={span}
-										totalDuration={totalDuration}
-										startTime={startTime}
-										depth={0}
-										height={height}
-										width={width}
-										zoom={zoom}
-										selectedSpan={selectedSpan}
-										setHoveredSpan={setHoveredSpan}
-										setSelectedSpan={setSelectedSpan}
-										setTooltipCoordinates={
-											setTooltipCoordinatesImpl
-										}
-									/>
-								)
-							})}
-						</svg>
-						{hoveredSpan && (
-							// TODO: Move tooltip component outside the graph to prevent rerenders.
-							<Box
-								position="fixed"
-								display="flex"
-								flexDirection="column"
-								gap="6"
-								style={{
-									left: tooltipCoordinates.x - 10,
-									top: tooltipCoordinates.y + 5,
-									backgroundColor: '#fff',
-									padding: '4px 8px',
-									borderRadius: '4px',
-									zIndex: 1000,
-								}}
-								shadow="small"
-								border="dividerWeak"
-							>
-								<Text weight="bold" lines="1">
-									{hoveredSpan.spanName}
-								</Text>
-								<Text lines="1">
-									Duration:{' '}
-									{getTraceDurationString(
-										hoveredSpan.duration,
-									)}
-								</Text>
-								<Text lines="1">
-									Start:{' '}
-									{getTraceDurationString(
-										hoveredSpan.start,
-									) || '0ms'}
-								</Text>
-							</Box>
-						)}
-					</Box>
-				</Box>
+				<TraceFlameGraph
+					trace={data.trace.trace}
+					errors={data.trace.errors}
+					hoveredSpan={hoveredSpan}
+					selectedSpan={selectedSpan}
+					startTime={startTime}
+					totalDuration={totalDuration}
+					onSpanSelect={setSelectedSpan}
+					onSpanMouseEnter={setHoveredSpan}
+				/>
 
 				<Box mt="12">
 					<Tabs<TraceTabs>
